// src/types.ts
export interface QuestionOption {
    text?: string;
    image?: string;
}

export interface Question {
  id: string;
  question: string;
  questionImage?: string;
  options: QuestionOption[];
  correctAnswer: number;
  category: string;
  explanation: string;
  created_at?: string;
}